requirements.txt
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
marshmallow==2.16.3
psycopg2==2.7.6.1
factory_boy==2.11.1
ipython
ipdb
